functions/requirements.txt
firebase-functions==0.1.0
firebase-admin==6.2.0
tensorflow>=2.20.0
keras>=3.0.0
numpy>=1.24.0
Pillow>=10.0.0

